refactor(editor): EditorPane 抽出 load/save/rename 三個 hook
PR-A 把載入狀態機、retry、canSave 語意、儲存 toast、recent files 重映射與
saveButtonRef 全疊在 EditorPane.tsx（732 行），codex 檔案體質 review 判定
必須在 PR-B 加功能之前先拆。

拆法（純搬移，行為零變更）：
- hooks/useEditorPaneLoadState.ts — 載入 effect、綁 buffer key 的 loadError、
  loadAttempt retry；LoadError 型別與 loadErrorMessage 一併內化
- hooks/useEditorSaveFlow.ts — handleSave、saveUntitledBuffer、儲存 toast、
  save button anchor fallback、readStatAfterWrite、saveErrorReason
- hooks/useEditorRenameFlow.ts — handleRenameSubmit、rename warning、
  recent files 重映射
- hooks/useRenamePopoverState.ts — 命名 popover 的四個 UI state 收攏成穩定的
  控制介面（openRename / openSave / close / setWarning），因為 save 與 rename
  兩條流程都要驅動它，否則要對兩個 hook 各傳四個 setter
- editor-pane-naming.ts — 純檔名／路徑 helper，讓 hook 不必反向 import
  EditorPane（會成 import cycle）

EditorPane 只留 store wiring、mode 推導、attach/cleanup/external-change effect
與 render：732 → 362 行。

測試一行未改（斷言、檔案、數量皆同）即為行為未變的證明：
4060 tests / 346 files 全綠，lint、build 亦綠。

註：useRenamePopoverState 與 useEditorPaneLoadState 各有一處
react-hooks/set-state-in-effect 抑制。該 pattern 原本就存在於 EditorPane，
只是 react-hooks compiler 對原本那個大 component 直接 bail out 所以不報；
搬進小 hook 後才被看見。本次維持原行為，不順手改寫。
