+++ spa/src/components/editor/EditorPane.tsx
@@ -1,12 +1,11 @@
 // spa/src/components/editor/EditorPane.tsx
-import { lazy, Suspense, useEffect, useCallback, useRef, useState } from 'react'
+import { lazy, Suspense, useEffect, useCallback } from 'react'
 import type { PaneRendererProps } from '../../lib/module-registry'
 import { useEditorStore } from '../../stores/useEditorStore'
 import { useEditorSettingsStore } from '../../stores/useEditorSettingsStore'
 import { useTabStore } from '../../stores/useTabStore'
 import { useWorkspaceStore } from '../../features/workspace/store'
 import { useI18nStore } from '../../stores/useI18nStore'
-import { getFsBackend } from '../../lib/fs-backend'
 import { openInAppFile } from '../../lib/open-in-app-file'
 import { MonacoWrapper } from './MonacoWrapper'
 import { DiffView } from './DiffView'
@@ -17,108 +16,18 @@ import { findPane } from '../../lib/pane-tree'
 import { bufferKey } from '../../lib/editor-buffer-key'
 import { STORAGE_ROOT } from '../../lib/storage-paths'
 import { createUniqueInAppFile } from '../../lib/inapp-namer'
-import { recordRecentFile } from '../../lib/recent-files/record-recent-file'
-import { useRecentFilesStore } from '../../stores/useRecentFilesStore'
-import { useUndoToast } from '../../stores/useUndoToast'
+import { getFsBackend } from '../../lib/fs-backend'
+import { displayName, isInvalidRename, isUntitledPath } from './editor-pane-naming'
+import { useRenamePopoverState } from './hooks/useRenamePopoverState'
+import { useEditorPaneLoadState } from './hooks/useEditorPaneLoadState'
+import { useEditorSaveFlow } from './hooks/useEditorSaveFlow'
+import { useEditorRenameFlow } from './hooks/useEditorRenameFlow'
 import type { FileSource } from '../../types/fs'
 import type { UntitledDocumentState } from '../../types/tab'
-import {
-  createMetadata,
-  untitledStoragePath,
-  untitledSuggestedName,
-} from '../../lib/editor-language'
 
 const TiptapEditor = lazy(() =>
   import('./TiptapEditor').then((m) => ({ default: m.TiptapEditor }))
 )
-
-function isUntitledPath(filePath: string): boolean {
-  return filePath.startsWith('untitled:')
-}
-
-function displayName(filePath: string, untitled?: UntitledDocumentState): string {
-  return untitled?.name ?? fileName(filePath)
-}
-
-function renamePath(filePath: string, nextName: string, untitled?: UntitledDocumentState): string {
-  return untitled ? `untitled:${nextName}` : siblingPath(filePath, nextName)
-}
-
-function fileName(filePath: string): string {
-  return filePath.split('/').pop() ?? filePath
-}
-
-function siblingPath(filePath: string, nextBaseName: string): string {
-  const separatorIndex = filePath.lastIndexOf('/')
-  return separatorIndex === -1 ? nextBaseName : `${filePath.slice(0, separatorIndex)}/${nextBaseName}`
-}
-
-function isInvalidRename(name: string): boolean {
-  const trimmed = name.trim()
-  return trimmed === '' || trimmed === '.' || trimmed === '..' || trimmed.includes('/') || trimmed.includes('\\')
-}
-
-function isCaseOnlyRename(oldPath: string, nextPath: string): boolean {
-  return oldPath !== nextPath && oldPath.toLowerCase() === nextPath.toLowerCase()
-}
-
-function renameWarningMessage(error: unknown): string {
-  if (error instanceof Error) {
-    if (/exist/i.test(error.message)) return 'File already exists'
-    return error.message || 'Rename failed'
-  }
-  return 'Rename failed'
-}
-
-// Spec 3.2 (T3.3): the failure toast has to say *why* the save failed, so the
-// reason is extracted from whatever the backend rejected with.
-function saveErrorReason(error: unknown): string {
-  if (error instanceof Error && error.message) return error.message
-  return String(error)
-}
-
-/**
- * Post-write stat, demoted to a best-effort refresh of `lastStat`.
- *
- * A successful `write` means the bytes are on disk — the save has happened. The
- * follow-up `stat` only exists to refresh the external-change baseline, so a
- * rejection there (flaky link, host just removed, permissions) must NOT be
- * reported as a failed save: the user would be told nothing was written while
- * the file on disk already carries their edits, and the buffer would stay dirty
- * (or, for an untitled document, stay unnamed and out of the recent list).
- *
- * `undefined` is a safe value for `markSaved`, which falls back to the buffer's
- * existing `lastStat`.
- */
-async function readStatAfterWrite(
-  backend: { stat: (path: string) => Promise<{ mtime: number; size: number }> },
-  path: string,
-): Promise<{ mtime: number; size: number } | undefined> {
-  try {
-    const stat = await backend.stat(path)
-    return { mtime: stat.mtime, size: stat.size }
-  } catch {
-    return undefined
-  }
-}
-
-// Spec 1.2: a load failure must never degrade into an empty buffer. Extract the
-// most specific reason we can so the error surface tells the user *why* the file
-// could not be read (empty string → the generic i18n fallback).
-function loadErrorMessage(error: unknown): string {
-  if (error instanceof Error) return error.message
-  if (typeof error === 'string') return error
-  return ''
-}
-
-/**
- * A pane-local load failure. Two shapes, because the two sources of failure carry
- * different text: `message` is a raw reason handed to us by the backend (already
- * a sentence, not translatable); `messageKey` is an i18n key for failures we
- * diagnose ourselves (spec 1.2b: no FS backend resolved at all). Keeping the key
- * unresolved until render means a locale switch retranslates the surface.
- */
-type LoadError = { message: string; messageKey?: undefined } | { messageKey: string; message?: undefined }
 
 function sourceIdentity(source: FileSource): string {
   return source.type === 'daemon' ? `daemon:${source.hostId}` : source.type
@@ -199,25 +108,21 @@ function EditorPaneInner({ paneId, source, filePath, untitled, isActive }: { pan
   // absent must not masquerade as modified — that was what made every remote
   // file look dirty the moment it opened.
   const canSave = buffer ? (buffer.isDirty || (!!buffer.untitled && !buffer.lastStat)) : false
-  // Per-pane load failure (spec 1.2). Local state on purpose: the store never
-  // learns about a failed load because no buffer is created for it.
-  //
-  // Tagged with the buffer key that produced it: a pane outlives a file switch,
-  // and the load effect only clears the error AFTER the commit — so an untagged
-  // error painted over the next file for a frame, offering a Retry button that
-  // would actually retry that next file.
-  const [loadError, setLoadError] = useState<{ key: string; error: LoadError } | null>(null)
-  const activeLoadError = loadError?.key === key ? loadError.error : null
-  const [loadAttempt, setLoadAttempt] = useState(0)
-  const [renameAnchorRect, setRenameAnchorRect] = useState<DOMRect | null>(null)
-  const [renameMode, setRenameMode] = useState<'rename' | 'save'>('rename')
-  const [renameInitialValue, setRenameInitialValue] = useState<string>()
-  const [renameWarning, setRenameWarning] = useState<string>()
-  // Fallback anchor for the naming popover. Whether the popover is NEEDED is a
-  // property of the buffer; WHERE it hangs is a property of the UI, and the two
-  // used to be conflated — a save that arrived without a rect (the keyboard
-  // path) was dropped entirely instead of anchoring itself to the Save button.
-  const saveButtonRef = useRef<HTMLButtonElement | null>(null)
+
+  const renamePopover = useRenamePopoverState(filePath)
+  const { activeLoadError, retryLoad } = useEditorPaneLoadState({
+    key, source, sourceId, filePath, untitled, isUntitled,
+  })
+  const { handleSave, saveUntitledBuffer, saveButtonRef } = useEditorSaveFlow({
+    key, source, filePath, paneId, untitled, popover: renamePopover.controls, t,
+  })
+  const { handleRenameSubmit } = useEditorRenameFlow({
+    key, source, filePath, currentName,
+    renameMode: renamePopover.mode,
+    saveUntitledBuffer,
+    popover: renamePopover.controls,
+    t,
+  })
 
   const handleCursorChange = useCallback((line: number, column: number) => {
     useEditorStore.getState().updateCursor(paneId, line, column)
@@ -232,63 +137,10 @@ function EditorPaneInner({ paneId, source, filePath, untitled, isActive }: { pan
   }, [paneId, key])
 
   useEffect(() => {
-    setRenameAnchorRect(null)
-    setRenameWarning(undefined)
-  }, [filePath])
-
-  useEffect(() => {
     if (!isMarkdown && editorMode !== 'raw') {
       useEditorStore.getState().setEditorMode(paneId, 'raw')
     }
   }, [editorMode, isMarkdown, paneId])
-
-  // Load file on mount, cleanup buffer on unmount.
-  // `loadAttempt` is a deps-only retry trigger (see handleRetryLoad).
-  useEffect(() => {
-    let stale = false
-    setLoadError(null)
-    if (useEditorStore.getState().buffers[key]) return // already loaded
-    // Only an untitled pane may open an empty buffer (spec 1.2): it has no file
-    // behind it yet. Everything else — including "not found" — becomes a load
-    // error, because a buffer opened over an unreadable file would silently
-    // truncate that file on the next save.
-    if (isUntitled) {
-      useEditorStore.getState().openBuffer(key, '', createMetadata(source, filePath, untitled))
-      return
-    }
-
-    // Spec 1.2b: no backend at all (a `local` source outside Electron, a daemon
-    // source whose host was removed) used to bail out silently and leave the pane
-    // spinning on "Loading…" forever. Same silent-failure class as 1.2, reached by
-    // a different route — so it surfaces through the same retryable error state.
-    const backend = getFsBackend(source)
-    if (!backend) {
-      setLoadError({ key, error: { messageKey: 'editor.load_error.no_backend' } })
-      return
-    }
-
-    backend.read(filePath)
-      .then((data) => {
-        if (stale) return
-        const text = new TextDecoder().decode(data)
-        const metadata = createMetadata(source, filePath, untitled)
-        return backend.stat(filePath).then((stat) => {
-          if (stale) return
-          useEditorStore.getState().openBuffer(key, text, metadata, { mtime: stat.mtime, size: stat.size })
-        })
-      })
-      .catch((error: unknown) => {
-        if (stale) return
-        setLoadError({ key, error: { message: loadErrorMessage(error) } })
-      })
-
-    return () => { stale = true }
-  }, [filePath, isUntitled, key, sourceId, source, untitled, loadAttempt])
-
-  const handleRetryLoad = useCallback(() => {
-    setLoadError(null)
-    setLoadAttempt((attempt) => attempt + 1)
-  }, [])
 
   // Cleanup pane state only when the pane is truly gone, not just hidden by tab switching.
   useEffect(() => {
@@ -338,219 +190,6 @@ function EditorPaneInner({ paneId, source, filePath, untitled, isActive }: { pan
     // eslint-disable-next-line react-hooks/exhaustive-deps -- only re-check on tab activation, not on source/filePath change
   }, [isActive, isUntitled, key, source])
 
-  // T3.3: one place that turns a save attempt into user-visible feedback. It
-  // reuses the existing bottom-centre toast (`useUndoToast` / `GlobalUndoToast`)
-  // rather than adding a second notification surface.
-  const showSaveToast = useCallback((messageKey: string, params?: Record<string, string>) => {
-    useUndoToast.getState().show(t(messageKey, params))
-  }, [t])
-
-  const saveUntitledBuffer = useCallback(async (name: string) => {
-    const buf = useEditorStore.getState().buffers[key]
-    const backend = getFsBackend(source)
-    // No buffer / not an untitled pane are preconditions, not outcomes: there is
-    // no document to report on.
-    if (!buf || !untitled) return
-    // A missing backend IS an outcome, and the same one T3.3-4b reports on the
-    // ordinary save path. Returning silently here left the user believing the
-    // first save of a new document had landed when nothing was written at all.
-    if (!backend) {
-      showSaveToast('editor.save.failed', { reason: t('editor.load_error.no_backend') })
-      return
-    }
-
-    const trimmedName = name.trim()
-    if (isInvalidRename(trimmedName)) {
-      setRenameWarning('Invalid file name')
-      return
-    }
-
-    const nextPath = untitledStoragePath(trimmedName)
-    const nextKey = bufferKey(source, nextPath)
-    if (nextKey !== key && useEditorStore.getState().buffers[nextKey]) {
-      setRenameWarning('File already exists')
-      return
-    }
-    // An OPEN buffer is not the only way the name can be taken: the file may
-    // already sit on the backend with nothing open on it, and the write below
-    // is a blind overwrite. Probe the backend the same way `handleRenameSubmit`
-    // does — a successful stat means the path is occupied and the first save of
-    // this document must NOT clobber it.
-    try {
-      await backend.stat(nextPath)
-      setRenameWarning('File already exists')
-      return
-    } catch {
-      // Missing target is the expected, writable case.
-    }
-
-    // The WRITE decides the outcome — see `readStatAfterWrite`.
-    try {
-      await backend.write(nextPath, new TextEncoder().encode(buf.content))
-    } catch (err) {
-      showSaveToast('editor.save.failed', { reason: saveErrorReason(err) })
-      return
-    }
-
-    const newStat = await readStatAfterWrite(backend, nextPath)
-    const nextMetadata = buf.languageSource === 'manual'
-      ? { language: buf.language, languageSource: 'manual' as const, untitled: undefined }
-      : { ...createMetadata(source, nextPath), untitled: undefined }
-    useTabStore.getState().renameEditorPanes(source, filePath, nextPath)
-    useEditorStore.getState().renameBuffer(key, nextKey, nextMetadata)
-    useEditorStore.getState().markSaved(nextKey, newStat)
-    recordRecentFile({ kind: 'editor', source, filePath: nextPath })
-    useEditorStore.getState().setShowDiff(paneId, false)
-    setRenameAnchorRect(null)
-    setRenameInitialValue(undefined)
-    setRenameWarning(undefined)
-    // Confirming the name IS a save outcome (the file now exists on disk), so
-    // it reports like any other save. Only *opening* the popover is silent.
-    showSaveToast('editor.save.saved', { name: fileName(nextPath) })
-  }, [filePath, key, paneId, showSaveToast, source, t, untitled])
-
-  const handleSave = useCallback(async (anchorRect?: DOMRect) => {
-    const buf = useEditorStore.getState().buffers[key]
-    // No buffer: the pane is showing the loading / load-error surface, so there
-    // is no document and no save attempt to report.
-    if (!buf) return
-    // T3.3: an untouched, already-saved buffer is an explicit "nothing to do"
-    // outcome. This branch used to swallow every ⌘S without a trace, which is
-    // what made the key feel dead.
-    if (!buf.isDirty && buf.lastStat) {
-      showSaveToast('editor.save.unchanged')
-      return
-    }
-    if (buf.untitled) {
-      if (!buf.untitled.hasBeenRenamed) {
-        // Opening the name popover is not a save outcome — no toast here; the
-        // one that follows the user's confirmation comes from saveUntitledBuffer.
-        //
-        // The editors (Monaco / Tiptap) call `onSave()` with no rect, so falling
-        // back to the Save button's own rect is what keeps ⌘S from being a
-        // no-op. The button is rendered by this same component whenever a buffer
-        // exists — i.e. whenever this branch is reachable — so the ref is set.
-        const anchor = anchorRect ?? saveButtonRef.current?.getBoundingClientRect()
-        if (!anchor) return
-        setRenameMode('save')
-        setRenameAnchorRect(anchor)
-        setRenameInitialValue(untitledSuggestedName(buf.untitled))
-        setRenameWarning(undefined)
-        return
-      }
-      await saveUntitledBuffer(buf.untitled.name)
-      return
-    }
-
-    const backend = getFsBackend(source)
-    // An unresolvable backend is a failed save, not a no-op: without this the
-    // content silently stays unsaved (same silent-failure class as 1.2b).
-    if (!backend) {
-      showSaveToast('editor.save.failed', { reason: t('editor.load_error.no_backend') })
-      return
-    }
-    // Only the WRITE decides success/failure (see `readStatAfterWrite`).
-    try {
-      await backend.write(filePath, new TextEncoder().encode(buf.content))
-    } catch (err) {
-      // Replaces a console.error the user could never see; the buffer stays
-      // dirty and unmarked, so the toast is the only signal they get.
-      showSaveToast('editor.save.failed', { reason: saveErrorReason(err) })
-      return
-    }
-
-    const newStat = await readStatAfterWrite(backend, filePath)
-    useEditorStore.getState().markSaved(key, newStat)
-    recordRecentFile({ kind: 'editor', source, filePath })
-    useEditorStore.getState().setShowDiff(paneId, false)
-    showSaveToast('editor.save.saved', { name: fileName(filePath) })
-  }, [filePath, key, paneId, saveUntitledBuffer, showSaveToast, source, t])
-
-  const handleRenameSubmit = useCallback(async (nextName: string) => {
-    const currentBuffer = useEditorStore.getState().buffers[key]
-    if (!currentBuffer) return
-
-    if (renameMode === 'save') {
-      await saveUntitledBuffer(nextName)
-      return
-    }
-
-    if (isInvalidRename(nextName)) {
-      setRenameWarning('Invalid file name')
-      return
-    }
-
-    if (nextName === currentName) {
-      setRenameAnchorRect(null)
-      setRenameInitialValue(undefined)
-      setRenameWarning(undefined)
-      return
-    }
-
-    const nextPath = renamePath(filePath, nextName, currentBuffer.untitled)
-    const nextKey = bufferKey(source, nextPath)
-    if (nextKey !== key && useEditorStore.getState().buffers[nextKey]) {
-      setRenameWarning('File already exists')
-      return
-    }
-
-    if (currentBuffer.untitled) {
-      const nextUntitled: UntitledDocumentState = {
-        ...currentBuffer.untitled,
-        name: nextName,
-        hasBeenRenamed: true,
-      }
-      useTabStore.getState().renameEditorPanes(source, filePath, nextPath, { untitled: nextUntitled })
-      const nextMetadata = currentBuffer.languageSource === 'manual'
-        ? { language: currentBuffer.language, languageSource: 'manual' as const, untitled: nextUntitled }
-        : createMetadata(source, nextPath, nextUntitled)
-      useEditorStore.getState().renameBuffer(key, nextKey, nextMetadata)
-      setRenameAnchorRect(null)
-      setRenameInitialValue(undefined)
-      setRenameWarning(undefined)
-      return
-    }
-
-    const backend = getFsBackend(source)
-    // Same silent-failure class as T1.2b: without a backend the rename cannot
-    // happen, and returning quietly dismissed the popover as if it had.
-    if (!backend) {
-      setRenameWarning(t('editor.load_error.no_backend'))
-      return
-    }
-
-    if (!isCaseOnlyRename(filePath, nextPath)) {
-      try {
-        await backend.stat(nextPath)
-        setRenameWarning('File already exists')
-        return
-      } catch {
-        // Missing target is expected.
-      }
-    }
-
-    try {
-      if (currentBuffer?.lastStat) {
-        await backend.rename(filePath, nextPath)
-      }
-      useTabStore.getState().renameEditorPanes(source, filePath, nextPath)
-      const nextMetadata = currentBuffer?.languageSource === 'manual'
-        ? { language: currentBuffer.language, languageSource: 'manual' as const }
-        : createMetadata(source, nextPath)
-      useEditorStore.getState().renameBuffer(key, nextKey, nextMetadata)
-      // T3.2: the in-editor rename is the third path-mutating call site (the
-      // other two go through `remapPanesUnder`), and the ONLY one a remote file
-      // can take — `source` carries the daemon host, so the remap stays scoped
-      // to that host's entries.
-      useRecentFilesStore.getState().renamePath(source, filePath, nextPath)
-      setRenameAnchorRect(null)
-      setRenameInitialValue(undefined)
-      setRenameWarning(undefined)
-    } catch (error) {
-      setRenameWarning(renameWarningMessage(error))
-    }
-  }, [currentName, filePath, key, renameMode, saveUntitledBuffer, source, t])
-
   if (!buffer) {
     // Spec 1.2: a failed load renders an explicit, retryable error instead of an
     // empty editor. No buffer exists here, so there is nothing to save over the
@@ -570,7 +209,7 @@ function EditorPaneInner({ paneId, source, filePath, untitled, isActive }: { pan
           <button
             type="button"
             data-testid="editor-load-error-retry"
-            onClick={handleRetryLoad}
+            onClick={retryLoad}
             className="mt-1 rounded-md bg-accent px-3 py-1.5 text-xs text-text-inverse hover:bg-accent-hover"
           >
             {t('editor.load_error.retry')}
@@ -593,12 +232,7 @@ function EditorPaneInner({ paneId, source, filePath, untitled, isActive }: { pan
         onSave={handleSave}
         saveButtonRef={saveButtonRef}
         onDiff={() => useEditorStore.getState().setShowDiff(paneId, !showDiff)}
-        onRenameStart={(anchorRect) => {
-          setRenameMode('rename')
-          setRenameAnchorRect(anchorRect)
-          setRenameInitialValue(undefined)
-          setRenameWarning(undefined)
-        }}
+        onRenameStart={(anchorRect) => renamePopover.controls.openRename(anchorRect)}
         onBufferSwitch={(newKey) => {
           // Dirty-guard (spec v1.3 §4.8): prompt before leaving a dirty pane.
           const currentKey = bufferKey({ type: 'inapp' }, filePath)
@@ -706,20 +340,16 @@ function EditorPaneInner({ paneId, source, filePath, untitled, isActive }: { pan
           useEditorStore.getState().setEditorMode(paneId, mode)
         }}
       />
-      {renameAnchorRect && (
+      {renamePopover.anchorRect && (
         <RenamePopover
-          anchorRect={renameAnchorRect}
+          anchorRect={renamePopover.anchorRect}
           currentName={currentName}
-          initialValue={renameInitialValue}
-          allowUnchangedSubmit={renameMode === 'save'}
+          initialValue={renamePopover.initialValue}
+          allowUnchangedSubmit={renamePopover.mode === 'save'}
           onConfirm={handleRenameSubmit}
-          onCancel={() => {
-            setRenameAnchorRect(null)
-            setRenameInitialValue(undefined)
-            setRenameWarning(undefined)
-          }}
-          error={renameWarning}
-          onClearError={() => setRenameWarning(undefined)}
+          onCancel={renamePopover.controls.close}
+          error={renamePopover.warning}
+          onClearError={() => renamePopover.controls.setWarning(undefined)}
           placeholder="File name"
           validateName={(trimmedDraft, currentName) => {
             if (!trimmedDraft || trimmedDraft === currentName) return undefined
